implemented gateway Logic

+++ Api-gateway/server.js
@@ -31,22 +31,28 @@ app.get('/', (req, res) => {
     res.status(200).send('Working');
 });
 
-const middleware = async (req, res,next) => {
+const middleware = async (req, res, next) => {
+    console.log('req: ', req.session);
     const { token } = req.session;
+    console.log('token: ', token);
     try {
+        if (!token) throw new Error('unauthorized');
+        console.log('unauthorized: ');
         const key = await client.get(token);
+        console.log('key: ', key);
         if (!key) throw new Error('unauthorized');
         const decoded = jwt.verify(key, process.env.SECRET);
+        console.log('decoded: ', decoded);
         if (decoded) {
             next();
         } else {
-            await client.del(token);
-            res.status(401).json({ error: 'Unauthorized' });
+            await client.del(token); 
+            res.redirect(401, process.env.LOGIN_URL);
         }
     } catch (error) {
-        await client.del(token);
+        if (token) await client.del(token);
         if (error.message === 'jwt expired' || error.message === 'unauthorized') {
-            res.status(401).json({ error: error.message });
+            res.redirect(401, process.env.LOGIN_URL);
         } else {
             res.status(500).json({ error: error.message });
         }
@@ -54,45 +60,57 @@ const middleware = async (req, res,next) => {
 }
 
 app.post('/login', async (req, res) => {
+    const { token } = req.session;
+    const { redirectUrl } = req.query;
+    const { email, password } = req.body;
     try {
-        const { email, password } = req.body;
-        const { redirectUrl } = req.query;
+        if (token) {
+            const key = await client.get(token);
+            if (key) {
+                const decoded = jwt.verify(key, process.env.SECRET);
+                if (decoded) {
+                    if (redirectUrl) {
+                        res.redirect(301, redirectUrl);
+                        return
+                    } else {
+                        res.redirect(301, process.env.HOMEPAGE_URL);
+                        return
+                    }
+                }
+            }
+        }
         if (!email || !password) {
             return res.status(400).json({ error: `${email ? 'Password required' : 'Email required'}` });
         }
-        const token = jwt.sign({ email }, process.env.SECRET, { expiresIn: '20s' });
+        const newToken = jwt.sign({ email }, process.env.SECRET, { expiresIn: '20s' });
         const key = uuid.v4();
-        const redisData = await client.setEx(key, 300000, token);
+        const redisData = await client.setEx(key, 300000, newToken);
         console.log(redisData);
-        req.session.token = {token};
+        req.session.token = key;
         if (redirectUrl) {
             res.redirect(301, redirectUrl);
+            return
         } else {
             res.redirect(301, process.env.HOMEPAGE_URL);
+            return
         }
     } catch (error) {
         res.status(500).json({ error: error.message });
     }
 });
 
-app.post('/redirect', async (req, res, next) => {
+app.post('/redirect', middleware, async (req, res, next) => {
     const { redirectUrl } = req.body;
-    const { token } = req.session;
+    console.log('redirectUrl: ', redirectUrl);
     try {
-        console.log(token);
-        const key = await client.get(token);
-        if (!key) throw new Error('unauthorized');
-        const decoded = jwt.verify(key, process.env.SECRET);
-        if (decoded) {
-            res.redirect(301, `${redirectUrl}?token=${token}`);
+        if (redirectUrl) {
+            res.redirect(301, redirectUrl);
         } else {
-            await client.del(token);
-            res.status(401).json({ error: 'Unauthorized' });
+            res.redirect(301, process.env.HOMEPAGE_URL);
         }
     } catch (err) {
-        await client.del(token);
-        if (err.message === 'jwt expired' || err.message === 'unauthorized') {
-            res.status(401).json({ error: err.message });
+        if (err.error === 'jwt expired' || err.error === 'unauthorized') {
+            res.redirect(401, process.env.LOGIN_URL);
         } else {
             res.status(500).json({ error: err.message });
         }
@@ -100,36 +118,35 @@ app.post('/redirect', async (req, res, next) => {
 });
 
 app.get('/logout', async (req, res) => {
-    const { token } = req.params;
+    const { token } = req.session;
     try {
         await client.del(token);
+        req.session.destroy();
         res.status(200).json({ message: 'Logged out successfully' });
     } catch (error) {
         res.status(500).json({ error: error.message });
     }
 });
 
-app.post('/verify', async (req, res) => {
-    const { token } = req.body;
-    try {
-        const key = await client.get(token);
-        if (!key) throw new Error('unauthorized');
-        const decoded = jwt.verify(key, process.env.SECRET);
-        if (decoded) {
-            next();
-        } else {
-            await client.del(token);
-            res.status(401).json({ error: 'Unauthorized' });
-        }
-    } catch (error) {
-        await client.del(token);
-        if (error.message === 'jwt expired' || error.message === 'unauthorized') {
-            res.status(401).json({ error: error.message });
-        } else {
-            res.status(500).json({ error: error.message });
-        }
-    }
-});
+// app.post('/verify', async (req, res) => {
+//     const { token } = req.session.token;
+//     try {
+//         const key = await client.get(token);
+//         if (!key) throw new Error('unauthorized');
+//         const decoded = jwt.verify(key, process.env.SECRET);
+//         if (decoded) {
+//             next();
+//         } else {
+//             res.status(401).json({ error: 'Unauthorized' });
+//         }
+//     } catch (error) {
+//         if (error.message === 'jwt expired' || error.message === 'unauthorized') {
+//             res.status(401).json({ error: error.message });
+//         } else {
+//             res.status(500).json({ error: error.message });
+//         }
+//     }
+// });
 
 app.listen(process.env.PORT, () => {
     console.log(`Server is running on port ${process.env.PORT}`);
